# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Food Delivery service API
  description: |-
    Toptal test project Food Delivery service API docs.
  version: '1.0'
servers:
  - url: http://localhost:3000
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Meal:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        title:
          type: string
          description: Meal title
          example: Chicken
        description:
          type: string
          description: Meal description
          example: Well done Chicken
        price:
          type: integer
          minimum: 1
        section:
          type: string
          description: 'Meal Section'
          enum: ['breakfast', 'lunch', 'dinner', 'appetizers', 'dessert']
        restaurant_uuid:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    UpdateMeal:
      type: object
      properties:
        title:
          type: string
          description: Meal title
          example: Chicken
        description:
          type: string
          description: Meal description
          example: Well done Chicken
        price:
          type: integer
          minimum: 1
        section:
          type: string
          description: 'Meal Section'
          enum: ['breakfast', 'lunch', 'dinner', 'appetizers', 'dessert']
...
    Coupon:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        coupon_code:
          type: string
          example: "SPECIAL20"
        percentage:
          type: number
          example: 20
        restaurant_uuid:
          type: string
          example: "232642a2-e8f5-475b-966e-1061578b555b"
        status:
          type: string
          example: "active"
        created_at:
          type: string
          example: "2024-04-30T12:00:00Z"
    CouponUpdate:
      type: object
      properties:
        percentage:
          type: integer
          description: New discount percentage
          example: 30
  responses:
    BadRequest:
      description: 'Bad Request'
    Forbidden:
      description: 'Forbidden'
    NotFound:
      description: 'Not Found'
paths:
  /registrations:
    post:
      tags: [Authentication]
      summary: 'Register as a new user'
      description: |-
        Response will contain a token. This token can be used through "Authorization: Bearer TOKEN_HERE"
        header to make authenticated requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minimum: 8
                  maximum: 50
                  example: 'my.V3ry_Secu?re!$Password'
                role:
                  type: string
                  enum: [customer, owner]
      responses:
        '201':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: object
                properties:
                  uuid:
                    type: string
                    description: 'User UUID'
                    format: uuid
                  email:
                    type: string
                    description: 'User email'
                    format: email
                  role:
                    type: string
                    description: 'User role'
                    enum: [customer, owner]
                  created_at:
                    type: string
                    format: date-time
                  token:
                    type: string
                    format: byte
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiVG9wdGFsIFRlc3QiLCJpYXQiOjE1MTYyMzkwMjJ9.HKiCUjqQtglzbAe0ARyow9DLf_qcJWkijzKm5eU7Dsk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Conflict - User with provided username or email already exists.
  /tokens:
    post:
      tags: [Authentication]
      summary: 'Create an authentication token'
      description: |-
        Response will contain a token. This token can be used through "Authorization: Bearer TOKEN_HERE"
        header to make authenticated requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  example: 'my.V3ry_Secu?re!$Password'
      responses:
        '200':
          description: 'Token created successfully'
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    format: byte
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiVG9wdGFsIFRlc3QiLCJpYXQiOjE1MTYyMzkwMjJ9.HKiCUjqQtglzbAe0ARyow9DLf_qcJWkijzKm5eU7Dsk'
        '401':
          description: Unauthorized - Invalid credentials.
...
  /meals/{uuid}:
    get:
      tags: [Meals]
      summary: Get a signle meal
      description: Get details of a specific meal by UUID.
      parameters:
        - in: path
          name: uuid
          required: true
          description: UUID of the meal to retrieve
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meal'
        '404':
          description: Not Found
    patch:
      tags: [Meals]
      summary: Update a single meal
      description: Update details of a specific meal by UUID. For Admin or meal restaurant owner.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeal'
      parameters:
        - name: uuid
          in: path
          description: Meal UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The meal was successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [Meals]
      summary: Delete a meal
      description: Delete a meal by UUID. For Admin or meal restaurant owner.
      parameters:
        - name: uuid
          in: path
          description: Meal UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: The meal was successfully deleted.
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
  /coupons/{uuid}:
    get:
      summary: Get a single coupon
      description: Get details of a specific coupon by UUID. For Admin or coupon restaurant owner.
      tags: 
        - Coupons
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coupon'
        '404':
          description: Not found
    patch:
      tags: 
        - Coupons
      summary: Update a single coupon
      description: Update details of a specific coupon by UUID. For Admin or coupon restaurant owner.
      parameters:
        - in: path
          name: uuid
          required: true
          description: UUID of the coupon to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CouponUpdate'
      responses:
        '200':
          description: Coupon updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coupon'
        '400':
          description: Bad request (e.g., validation error)
        '404':
          description: Coupon not found or already inactive
    delete:
      summary: Delete a coupon
      description: Delete a coupon by UUID. For Admin or coupon restaurant owner.
      tags: 
        - Coupons
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Successful operation
        '404':
          description: Not found